+++ seneca.js
@@ -250,7 +250,8 @@ const seneca_util = {
   pattern: Common.pattern,
   print: Common.print,
   error: error,
-
+  deep: Common.deepextend,
+  
   // Legacy
   deepextend: Common.deepextend,
   recurse: Common.recurse,
@@ -403,8 +404,10 @@ function make_seneca(initial_options) {
   private$.actnid = Nid({ length: opts.$.idlen })
   private$.didnid = Nid({ length: opts.$.didlen })
 
+  // Instance specific incrementing counters to create unique function names
   var next_action_id = Common.autoincr()
-
+  var next_ready_id = Common.autoincr()
+  
   // These need to come from options as required during construction.
   opts.$.internal.actrouter = opts.$.internal.actrouter || Patrun({ gex: true })
   opts.$.internal.subrouter = opts.$.internal.subrouter || Patrun({ gex: true })
@@ -687,8 +690,10 @@ function make_seneca(initial_options) {
       actdef.args = Common.deepextend(pattern)
     }
 
-    actdef.id = action.name + '_' + next_action_id()
-    actdef.name = action.name
+    var action_name = (null == action.name || '' === action.name) ?
+        'action' : action.name
+    actdef.id = action_name + '_' + next_action_id()
+    actdef.name = action_name
     actdef.func = action
 
     // Canonical string form of the action pattern.
@@ -917,10 +922,23 @@ function make_seneca(initial_options) {
 
     if ('function' === typeof ready) {
       setImmediate(function register_ready() {
+
+        var ready_call = function() {
+          ready.call(self)
+        }
+
+        var ready_name = ( (
+          null == ready.name ||
+            '' === ready.name ||
+            'ready' === ready.name
+        ) ? 'ready_' : ready.name + '_ready_' ) + next_ready_id()
+        
+        Object.defineProperty(ready_call, "name", { value: ready_name })
+
         if (root$.private$.ge.isclear()) {
-          execute_ready(ready.bind(self))
+          execute_ready(self, ready_call)
         } else {
-          root$.private$.ready_list.push(ready.bind(self))
+          root$.private$.ready_list.push(ready_call)
         }
       })
     }
@@ -1092,19 +1110,20 @@ function make_seneca(initial_options) {
   // if the task queue is emptied.
   function action_queue_clear() {
     root$.emit('ready')
-    execute_ready(root$.private$.ready_list.shift())
+    execute_ready(root$, root$.private$.ready_list.shift())
 
     if (root$.private$.ge.isclear()) {
       while (0 < root$.private$.ready_list.length) {
-        execute_ready(root$.private$.ready_list.shift())
+        execute_ready(root$, root$.private$.ready_list.shift())
       }
     }
   }
 
-  function execute_ready(ready_func) {
+  function execute_ready(instance, ready_func) {
     if (null == ready_func) return
 
     try {
+      instance.log.debug({kind:'ready', case:'call', name: ready_func.name})
       ready_func()
     } catch (ready_err) {
       var err = error(ready_err, 'ready_failed', { message: ready_err.message })